+++ TagCloud/PointFExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Drawing;
 
-namespace TagsCloudVisualization
+namespace TagCloud
 {
     public static class PointFExtensions
     {
