+++ TagCloud/TagCloud.csproj
@@ -1,0 +1,22 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net48</TargetFramework>
+        <LangVersion>latest</LangVersion>
+        <OutputType>Exe</OutputType>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Autofac" Version="6.3.0" />
+      <PackageReference Include="CommandLineParser" Version="2.9.0-preview1" />
+      <PackageReference Include="DeepMorphy" Version="2.0.3" />
+      <PackageReference Include="FluentAssertions" Version="6.2.0" />
+      <PackageReference Include="NUnit" Version="3.13.2" />
+      <PackageReference Include="WeCantSpell.Hunspell" Version="3.0.1" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="ExamplesIMG" />
+    </ItemGroup>
+
+</Project>
