+++ TagCloud/Templates/WordParameter.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing;
 
-namespace TagsCloudVisualization
+namespace TagCloud.Templates
 {
     public class WordParameter
     {
