+++ TagCloud/TextHandlers/Converters/IWordConverter.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace TagCloud.TextHandlers
+namespace TagCloud.TextHandlers.Converters
 {
     public interface IWordConverter
     {
