+++ TagCloud/TextHandlers/Parser/WordsParser.cs
@@ -1,16 +1,10 @@
 ﻿using System.Collections.Generic;
 using System.IO;
-using TagsCloudVisualization;
 
 namespace TagCloud.TextHandlers.Parser
 {
     public class WordsParser : ITextParser
     {
-
-        public WordsParser(string path)
-        {
-        }
-
         public IEnumerable<string> GetWords(string path)
         {
             return File.ReadAllLines(path);
